优化页面展示

+++ code/app/model/WebScanModel.php
@@ -20,13 +20,13 @@ class WebScanModel extends BaseModel
         while (true) {
             processSleep(1);
             $endTime = date('Y-m-d', time() - 86400 * 15);
-            $where[] = ['is_delete','=',0];
-            $where[] = ['status','=',1];
+            $where[] = ['is_delete', '=', 0];
+            $where[] = ['status', '=', 1];
             $list = self::getAppStayScanList('crawler_time');
             $count = Db::table('app')->whereTime('crawler_time', '<=', $endTime)->where($where)->count('id');
             print("开始执行rad扫描任务,{$count} 个项目等待扫描..." . PHP_EOL);
             foreach ($list as $value) {
-                if (!self::checkToolAuth(1,$value['id'],'rad')) {
+                if (!self::checkToolAuth(1, $value['id'], 'rad')) {
                     continue;
                 }
 
@@ -90,15 +90,15 @@ class WebScanModel extends BaseModel
                 }*/
                 $urlList = json_decode(file_get_contents($pathArr['tool_result']), true);
                 foreach ($urlList as $val) {
-                    $val['URL'] = rtrim($val['URL'],'/');
+                    $val['URL'] = rtrim($val['URL'], '/');
                     $arr = parse_url($val['URL']);
-                    $blackExt = ['.js', '.css', '.png', '.jpg', '.jpeg', '.gif', '.mp3', '.mp4','.ico','.bmp','.wmv','.avi','.psd'];
+                    $blackExt = ['.js', '.css', '.png', '.jpg', '.jpeg', '.gif', '.mp3', '.mp4', '.ico', '.bmp', '.wmv', '.avi', '.psd'];
                     //if (!isset($arr['query']) or (isset($arr['path']) && in_array_strpos($arr['path'], $blackExt)) or (strpos($arr['query'], '=') === false)) {
-                    if (isset($arr['path']) && in_array_strpos(strtolower($arr['path']), $blackExt) || in_array_strpos(strtolower($val['URL']),$blackExt)) {
+                    if (isset($arr['path']) && in_array_strpos(strtolower($arr['path']), $blackExt) || in_array_strpos(strtolower($val['URL']), $blackExt)) {
                         addlog(["rad扫描跳过无意义URL", $val['URL']]);
                         continue;
                     }
-                    if (!Db::name('urls')->where('hash',md5($val['URL']))->count()) {
+                    if (!Db::name('urls')->where('hash', md5($val['URL']))->count()) {
                         $newData = [
                             'app_id' => $id,
                             'method' => $val['Method'],
@@ -114,11 +114,38 @@ class WebScanModel extends BaseModel
                         addlog(["rad扫描数据写入成功", json_encode($newData)]);
                     }
                 }
-                PluginModel::addScanLog($value['id'], __METHOD__, 0, 1,1, ['content' => $urlList]);
+                PluginModel::addScanLog($value['id'], __METHOD__, 0, 1, 1, ['content' => $urlList]);
             }
             sleep(30);
         }
 
+    }
+
+    public static function editXrayYaml($filename)
+    {
+        $arr = @yaml_parse_file($filename);
+        if ($arr) {
+            $arr['http']['proxy_rule'][0]['match'] = '*';
+            $proxyArr = Db::name('proxy')->where('status', 1)->limit(3)->orderRand()->select()->toArray();
+            $proxy = [];
+            foreach ($proxyArr as $v) {
+                $result = testAgent($v['host'], $v['port']);
+                if ($result == 200) {
+                    $proxy[] = 'http://' . $v['host'] . ":{$v['port']}";
+                }
+            }
+            $arr['http']['proxy_rule'][0]['servers'] = [];
+            if ($proxy) {
+                $weight = random_split(10, count($proxy));
+                foreach ($proxy as $k => $v) {
+                    $arr['http']['proxy_rule'][0]['servers'][] = [
+                        'addr' => $v,
+                        'weight' => $weight[$k],
+                    ];
+                }
+            }
+            yaml_emit_file($filename, $arr);
+        }
     }
 
     public static function xray()
@@ -126,13 +153,13 @@ class WebScanModel extends BaseModel
         while (true) {
             processSleep(1);
             $endTime = date('Y-m-d', time() - 86400 * 15);
-            $where[] = ['is_delete','=',0];
-            $where[] = ['status','=',1];
+            $where[] = ['is_delete', '=', 0];
+            $where[] = ['status', '=', 1];
             $list = self::getAppStayScanList('xray_scan_time');
             $count = Db::table('app')->whereTime('xray_scan_time', '<=', $endTime)->where($where)->count('id');
             print("开始执行xray漏洞扫描任务,{$count} 个项目等待扫描..." . PHP_EOL);
             foreach ($list as $val) {
-                if (!self::checkToolAuth(1,$val['id'],'xray')) {
+                if (!self::checkToolAuth(1, $val['id'], 'xray')) {
                     continue;
                 }
 
@@ -146,42 +173,15 @@ class WebScanModel extends BaseModel
                 $pathArr = getSavePath($url, "xray", $id);
 
                 //初始化清理目录
-                if (file_exists($pathArr['tool_result'])) {
-                    unlink($pathArr['tool_result']);
-                }
-                if (file_exists($pathArr['cmd_result'])) {
-                    unlink($pathArr['cmd_result']);
-                }
+                if (file_exists($pathArr['tool_result'])) unlink($pathArr['tool_result']);
+                if (file_exists($pathArr['cmd_result'])) unlink($pathArr['cmd_result']);
+
 
                 if (file_exists($pathArr['tool_result']) == false) {
                     // 设置代理
                     $filename = '/data/tools/xray/config.yaml';
                     if (!$val['is_intranet']) {  // 不是内网
-                        if (file_exists($filename)) {
-                            $arr = @yaml_parse_file($filename);
-                            if ($arr) {
-                                $arr['http']['proxy_rule'][0]['match'] = '*';
-                                $proxyArr = Db::name('proxy')->where('status', 1)->limit(3)->orderRand()->select()->toArray();
-                                $proxy = [];
-                                foreach ($proxyArr as $v) {
-                                    $result = testAgent($v['host'], $v['port']);
-                                    if ($result == 200) {
-                                        $proxy[] = 'http://' . $v['host'] . ":{$v['port']}";
-                                    }
-                                }
-                                $arr['http']['proxy_rule'][0]['servers'] = [];
-                                if ($proxy) {
-                                    $weight = random_split(10, count($proxy));
-                                    foreach ($proxy as $k => $v) {
-                                        $arr['http']['proxy_rule'][0]['servers'][] = [
-                                            'addr' => $v,
-                                            'weight' => $weight[$k],
-                                        ];
-                                    }
-                                }
-                                yaml_emit_file($filename, $arr);
-                            }
-                        }
+                        if (file_exists($filename)) self::editXrayYaml($filename);
                     } else {
                         @unlink($filename);
                     }
@@ -194,7 +194,7 @@ class WebScanModel extends BaseModel
                     $result = file_put_contents($pathArr['cmd_result'], $result);
                     if ($result == false) {
                         addlog(["xray写入执行结果失败", base64_encode($pathArr['cmd_result'])]);
-                        PluginModel::addScanLog($val['id'], __METHOD__, 0,2);
+                        PluginModel::addScanLog($val['id'], __METHOD__, 0, 2);
                         continue;
                     }
                 } else {
@@ -204,7 +204,7 @@ class WebScanModel extends BaseModel
                 if (file_exists($pathArr['tool_result']) == false) {
                     addlog("xray扫描结果文件不存在:{$pathArr['tool_result']},扫描URL失败: {$url}");
                     Db::table('app')->where(['id' => $id])->save(['xray_scan_time' => date('2048-m-d H:i:s')]);
-                    PluginModel::addScanLog($val['id'], __METHOD__, 0,2);
+                    PluginModel::addScanLog($val['id'], __METHOD__, 0, 2);
                     continue;
                 }
 
@@ -215,19 +215,19 @@ class WebScanModel extends BaseModel
                         'app_id' => $val['id'],
                         'create_time' => substr($value['create_time'], 0, 10),
                         'detail' => base64_encode($value['detail']),
-                        'plugin' => json_encode($value['plugin'],JSON_UNESCAPED_UNICODE),
-                        'target' => json_encode($value['target'],JSON_UNESCAPED_UNICODE),
+                        'plugin' => json_encode($value['plugin'], JSON_UNESCAPED_UNICODE),
+                        'target' => json_encode($value['target'], JSON_UNESCAPED_UNICODE),
                         'url' => $value['detail']['addr'],
                         'url_id' => $val['id'],
                         'user_id' => $user_id,
                         'poc' => $value['detail']['payload']
                     ];
                     $addr[] = $newData;
-                    echo "xray添加漏洞结果:" . json_encode($newData,JSON_UNESCAPED_UNICODE) . PHP_EOL;
+                    echo "xray添加漏洞结果:" . json_encode($newData, JSON_UNESCAPED_UNICODE) . PHP_EOL;
                     XrayModel::addXray($newData);
                 }
-                addlog(["xray扫描数据写入成功:" . json_encode($addr,JSON_UNESCAPED_UNICODE)]);
-                PluginModel::addScanLog($val['id'], __METHOD__, 0,1);
+                addlog(["xray扫描数据写入成功:" . json_encode($addr, JSON_UNESCAPED_UNICODE)]);
+                PluginModel::addScanLog($val['id'], __METHOD__, 0, 1);
             }
             sleep(30);
         }
@@ -249,7 +249,7 @@ class WebScanModel extends BaseModel
                 execLog($cmd, $result);
                 // 如果返回值长度是0说明任务没有执行
                 if (count($result) == 0) {
-                    PluginModel::addScanLog($v['id'], __METHOD__, 0,2);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 2);
                     Db::name('app')->where('id', $v['id'])->update(['agent_start_up' => 1, 'agent_time' => date('Y-m-d H:i:s', time())]);
 
                     $cmd = "cd {$agent} && nohup ./xray_linux_amd64 webscan --listen 0.0.0.0:{$v['xray_agent_port']} --json-output {$v['id']}.json   xray_{$v['id']}_{$v['xray_agent_port']} >> /dev/null 2>&1";
@@ -257,7 +257,7 @@ class WebScanModel extends BaseModel
                     systemLog($cmd);
                     addlog(["xray代理模式启动", json_encode($cmd)]);
                 }
-                PluginModel::addScanLog($v['id'], __METHOD__, 0,1);
+                PluginModel::addScanLog($v['id'], __METHOD__, 0, 1);
             }
             sleep(30);
         }
@@ -311,7 +311,7 @@ class WebScanModel extends BaseModel
             processSleep(1);
             $list = self::getAppStayScanList('nuclei_scan_time');
             foreach ($list as $v) {
-                if (!self::checkToolAuth(1,$v['id'],'nuclei')) {
+                if (!self::checkToolAuth(1, $v['id'], 'nuclei')) {
                     continue;
                 }
 
@@ -322,7 +322,7 @@ class WebScanModel extends BaseModel
                 $cmd = "cd $agent && ./nuclei -u {$v['url']} -json -o {$filename}";
                 systemLog($cmd);
                 if (!file_exists($filename)) {
-                    PluginModel::addScanLog($v['id'], __METHOD__, 0,2);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 2);
                     addlog(["nucel扫描失败，url:{$v['url']}"]);
                     continue;
                 }
@@ -364,12 +364,12 @@ class WebScanModel extends BaseModel
                 }
                 fclose($file);
                 if (!$temp) {
-                    PluginModel::addScanLog($v['id'], __METHOD__,0, 1);
-                    addlog(["nuclei扫描未发现漏洞:{$v['url']}，数据结构：".json_encode($temp)]);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 1);
+                    addlog(["nuclei扫描未发现漏洞:{$v['url']}，数据结构：" . json_encode($temp)]);
                     continue;
                 }
                 addlog(["nuclei扫描数据写入成功:" . json_encode($temp)]);
-                PluginModel::addScanLog($v['id'], __METHOD__, 0,1);
+                PluginModel::addScanLog($v['id'], __METHOD__, 0, 1);
             }
             sleep(120);
         }
@@ -383,7 +383,7 @@ class WebScanModel extends BaseModel
             processSleep(1);
             $list = self::getAppStayScanList('vulmap_scan_time');
             foreach ($list as $v) {
-                if (!self::checkToolAuth(1,$v['id'],'vulmap')) {
+                if (!self::checkToolAuth(1, $v['id'], 'vulmap')) {
                     continue;
                 }
 
@@ -395,13 +395,13 @@ class WebScanModel extends BaseModel
                 $cmd = "cd $agent && python3 vulmap.py -u {$v['url']} --output-json {$filename}";
                 systemLog($cmd);
                 if (!file_exists($filename)) {
-                    PluginModel::addScanLog($v['id'], __METHOD__, 0,1);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 1);
                     addlog(["vulmap扫描完成,没有发现漏洞，url:{$v['url']}"]);
                     continue;
                 }
                 $arr = json_decode(file_get_contents($filename), true);
                 if (!$arr) {
-                    PluginModel::addScanLog($v['id'], __METHOD__, 0,2);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 2);
                     addlog(["{$v['url']}文件内容不存在:{$filename}"]);
                     continue;
                 }
@@ -425,10 +425,10 @@ class WebScanModel extends BaseModel
                     ];
                     if (!Db::name('app_vulmap')->insert($data)) {
                         addlog(["app_vulmap数据写入失败:" . json_encode($data)]);
-                        PluginModel::addScanLog($v['id'], __METHOD__, 0, 2, 1,['content' => 'app_vulmap数据写入失败']);
+                        PluginModel::addScanLog($v['id'], __METHOD__, 0, 2, 1, ['content' => 'app_vulmap数据写入失败']);
                     };
                 }
-                PluginModel::addScanLog($v['id'], __METHOD__, 0,1);
+                PluginModel::addScanLog($v['id'], __METHOD__, 0, 1);
             }
 
             sleep(120);
@@ -444,7 +444,7 @@ class WebScanModel extends BaseModel
             processSleep(1);
             $list = self::getAppStayScanList('crawlergo_scan_time');
             foreach ($list as $val) {
-                if (!self::checkToolAuth(1,$val['id'],'crawlergo')) {
+                if (!self::checkToolAuth(1, $val['id'], 'crawlergo')) {
                     continue;
                 }
 
@@ -457,7 +457,7 @@ class WebScanModel extends BaseModel
                 $cmd = "cd $tools && ./cmd/crawlergo/crawlergo_cmd -c /usr/bin/google-chrome -o none --output-json $filename -f 'strict' -t 10 {$val['url']}";
                 systemLog($cmd);
                 if (!file_exists($filename)) {
-                    PluginModel::addScanLog($val['id'], __METHOD__,0, 2);
+                    PluginModel::addScanLog($val['id'], __METHOD__, 0, 2);
                     addlog(["crawlergo扫描失败，url:{$val['url']}"]);
                     continue;
                 }
@@ -483,7 +483,7 @@ class WebScanModel extends BaseModel
                 if ($data) {
                     Db::name('app_crawlergo')->insertAll($data);
                 }
-                PluginModel::addScanLog($val['id'], __METHOD__,0, 1);
+                PluginModel::addScanLog($val['id'], __METHOD__, 0, 1);
             }
             sleep(120);
         }
@@ -498,7 +498,7 @@ class WebScanModel extends BaseModel
             processSleep(1);
             $list = self::getAppStayScanList('dismap_scan_time');
             foreach ($list as $v) {
-                if (!self::checkToolAuth(1,$v['id'],'dismap')) {
+                if (!self::checkToolAuth(1, $v['id'], 'dismap')) {
                     continue;
                 }
 
@@ -509,7 +509,7 @@ class WebScanModel extends BaseModel
                 $cmd = "cd $tools && ./dismap -url {$v['url']} -output dismap.txt";
                 systemLog($cmd);
                 if (!file_exists($filename)) {
-                    PluginModel::addScanLog($v['id'], __METHOD__, 0,2);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 2);
                     addlog(["dismap扫描失败，url:{$v['url']}"]);
                     continue;
                 }
@@ -520,7 +520,7 @@ class WebScanModel extends BaseModel
                 while (!feof($file)) {
                     $result = fgets($file);
                     if (empty($result)) {
-                        PluginModel::addScanLog($v['id'], __METHOD__, 0,2);
+                        PluginModel::addScanLog($v['id'], __METHOD__, 0, 2);
                         addlog(["dismap 扫描目标结果为空", $v['url']]);
                         continue;
                     }
@@ -538,15 +538,15 @@ class WebScanModel extends BaseModel
                 //关闭被打开的文件
                 fclose($file);
                 if (!$data) {
-                    PluginModel::addScanLog($v['id'], __METHOD__,0, 2);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 2);
                     addlog(["dismap扫描数据不存在，url:{$v['url']}"]);
                     continue;
                 }
                 if (!Db::name('app_dismap')->insertAll($data)) {
-                    PluginModel::addScanLog($v['id'], __METHOD__, 0,2);
+                    PluginModel::addScanLog($v['id'], __METHOD__, 0, 2);
                     addlog(["app_dismap数据写入失败:" . json_encode($data)]);
                 };
-                PluginModel::addScanLog($v['id'], __METHOD__,0, 1);
+                PluginModel::addScanLog($v['id'], __METHOD__, 0, 1);
             }
 
             sleep(60);
